Organize docker controllers and models
