+++ back/routes/api.php
@@ -55,26 +55,26 @@
     });
 
     Route::prefix('images')->group(function () {
-        Route::get('/', [\App\Http\Controllers\ImagesController::class, 'index']);
-        Route::post('/', [\App\Http\Controllers\ImagesController::class, 'store']);
-        Route::put('/', [\App\Http\Controllers\ImagesController::class, 'update']);
-        Route::delete('/', [\App\Http\Controllers\ImagesController::class, 'destroy']);
+        Route::get('/', [\App\Http\Controllers\Docker\ImagesController::class, 'index']);
+        Route::post('/', [\App\Http\Controllers\Docker\ImagesController::class, 'store']);
+        Route::put('/', [\App\Http\Controllers\Docker\ImagesController::class, 'update']);
+        Route::delete('/', [\App\Http\Controllers\Docker\ImagesController::class, 'destroy']);
     });
 
     Route::prefix('instances')->group(function () {
-        Route::get('/', [\App\Http\Controllers\InstancesController::class, 'index']);
-        Route::post('/', [\App\Http\Controllers\InstancesController::class, 'store']);
-        Route::put('/', [\App\Http\Controllers\InstancesController::class, 'update']);
-        Route::delete('/', [\App\Http\Controllers\InstancesController::class, 'destroy']);
+        Route::get('/', [\App\Http\Controllers\Docker\InstancesController::class, 'index']);
+        Route::post('/', [\App\Http\Controllers\Docker\InstancesController::class, 'store']);
+        Route::put('/', [\App\Http\Controllers\Docker\InstancesController::class, 'update']);
+        Route::delete('/', [\App\Http\Controllers\Docker\InstancesController::class, 'destroy']);
     });
 
     Route::prefix('containers')->group(function () {
-        Route::post('/', [\App\Http\Controllers\ContainersController::class, 'store'])->middleware('jwtcheck');
-        Route::post('/{id}', [\App\Http\Controllers\ContainersController::class, 'action']);
-        Route::get('/{id}', [\App\Http\Controllers\ContainersController::class, 'get']);
-        Route::get('/{id}/logs', [\App\Http\Controllers\ContainersController::class, 'logs']);
-        Route::get('/{id}/inspect', [\App\Http\Controllers\ContainersController::class, 'inspect']);
-        Route::get('/{id}/binds', [\App\Http\Controllers\ContainersController::class, 'binds']);
+        Route::post('/', [\App\Http\Controllers\Docker\ContainersController::class, 'create']);
+        Route::post('/{id}', [\App\Http\Controllers\Docker\ContainersController::class, 'action']);
+        Route::get('/{id}', [\App\Http\Controllers\Docker\ContainersController::class, 'get']);
+        Route::get('/{id}/logs', [\App\Http\Controllers\Docker\ContainersController::class, 'logs']);
+        Route::get('/{id}/inspect', [\App\Http\Controllers\Docker\ContainersController::class, 'inspect']);
+        Route::get('/{id}/binds', [\App\Http\Controllers\Docker\ContainersController::class, 'binds']);
     });
 
 ?>
